requirements.txt
paramiko
PyYAML
pydantic
